display is wired up
