+++ index.html
@@ -2,7 +2,7 @@
 <html>
 <head>
   <meta charset="utf-8">
-  <title>React Time</title>
+  <title>Weatherly2</title>
 </head>
 <body>
   <div id="app"></div>
