+++ lib/10Day.js
@@ -1,13 +1,13 @@
-import '../styles/7Hour.css';
+import '../styles/10Day.css';
 import React from 'react';
 import Card from './Card';
 
-function SevenHourForecast (props) {
+function TenDay (props) {
   return (
-    <div className="SevenHourForecast">
-
+    <div className="TenDayForecast">
+      <h1>TenDay</h1>
     </div>
   );
 }
 
-export default SevenHour;
+export default TenDay;
