+++ lib/7Hour.js
@@ -1,13 +1,13 @@
-import '../styles/10Day.css';
+import '../styles/7Hour.css';
 import React from 'react';
 import Card from './Card';
 
-function TenDay (props) {
+function SevenHour (props) {
   return (
-    <div className="TenDayForecast">
-
+    <div className="SevenHourForecast">
+      <h1>SevenHour</h1>
     </div>
   );
 }
 
-export default TenDay;
+export default SevenHour;
