+++ lib/App.js
@@ -1,7 +1,7 @@
 import '../styles/App.css';
 import React, { Component } from 'react';
-import SevenHourForecast from './7Hour';
-import TenDayForecast from './10Day';
+import SevenHour from './7Hour';
+import TenDay from './10Day';
 import CurrentWeather from './CurrentWeather';
 import Welcome from './Welcome';
 
@@ -17,7 +17,14 @@ class App extends Component {
   }
 
   render() {
-
+    return (
+      <div className="App">
+        <CurrentWeather />
+        <SevenHour />
+        <TenDay />
+        <Welcome />
+      </div>
+    )
   }
 }
