+++ lib/Card.js
@@ -5,9 +5,7 @@ function Card (props) {
   
   return (
     <div className="Card">
-      {
-
-      }
+      <h1>Card</h1>
     </div>
   );
 }
