+++ lib/CurrentWeather.js
@@ -5,7 +5,7 @@ function CurrentWeather (props) {
 
   return (
     <div className="CurrentWeather">
-
+      <h1>CurrentWeather</h1>
     </div>
   );
 }
