+++ lib/Welcome.js
@@ -1,11 +1,10 @@
 import '../styles/Welcome.css';
 import React from 'react';
-import Search from './Search';
 
 function Welcome (props) {
   return (
     <div className="Welcome">
-
+      <h1>Welcome</h1>
     </div>
   );
 }
